requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
elasticsearch==7.6.0
idna==2.9
numpy==1.18.2
pandas==1.0.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
tika==1.24
urllib3==1.25.8
xlrd==1.2.0
